+++ noodles-vcf/CHANGELOG.md
@@ -41,7 +41,7 @@
 
 ### Changed
 
-  * vcf/header/parser: Return error on duplicate info ID.
+  * vcf/header/parser: Return error on duplicate info and filter ID.
 
   * vcf/reader/builder: Change `Builder::build_from_reader` to accept
     implementations of `Read`.